chore: rename verbs/cm libraries to ibverbs/rdmacm
Rename libraries to align with rdma-core naming convention:
- verbs -> ibverbs
- cm -> rdmacm
