+++ src/lib.rs
@@ -1,2 +1,2 @@
-pub mod cm;
-pub mod verbs;
+pub mod ibverbs;
+pub mod rdmacm;
