+++ tests/post_send_guard/one_guard_has_only_one_wr.rs
@@ -1,4 +1,4 @@
-use sideway::verbs::{
+use sideway::ibverbs::{
     address::AddressHandleAttribute,
     device,
     device_context::Mtu,
